==> code/7_plots_paper.R <==
#### Plots and tables
# Load packages
library(tidyverse)
library(data.table)
library(lubridate)
library(readxl)
library(wesanderson)
library(ggpubr)
library(zoo)
library(scales)
library(ggsci)
library(extrafont)
library(patchwork)
loadfonts(device = "win")

setwd("/media/fabe4028/suhome/Documents/GitHub/nowcasting_covid19/code")

# Import COVID data and Nowcast results
dat <- read_csv("../data/covid_deaths.csv")
# res_df <- read_csv("../results/results_20220421.csv") %>% filter(date >= "2020-10-20", date <="2021-05-21")
res_df <- read_csv("../results/results_20220419.csv") %>% filter(date >= "2020-10-20", date <= "2021-05-21")

# dat %>% filter(death_date >= "2020-10-20", death_date <="2021-05-21") %>% #nrow()
# mutate(delay = rep_date -death_date) %>% filter(delay > 35) %>% nrow()

# Plot theme and color
theme_set(theme_bw())
# wes_cols <- c(wes_palette("Darjeeling1", 5), wes_palette("GrandBudapest2"),wes_palette("Cavalcanti1"))
cols <- pal_nejm("default", alpha = 1)(8)

# Figure 1, observed and unreported.
now <- "2022-02-01"
start <- "2022-01-01"
dat_mod <- dat %>% filter(rep_date <= now)
obs_plot <- dat %>%
  group_by(date = death_date) %>%
  summarise(n_true_retro = n()) %>%
  left_join(dat_mod %>% group_by(date = death_date) %>%
    summarise(n_obs = n())) %>%
  mutate_if(is.integer, ~ replace(., is.na(.), 0)) %>%
  filter(date >= start, date <= now) %>%
  ggplot() +
  geom_col(aes(date, n_true_retro, fill = "Occurred but not yet reported")) +
  geom_col(aes(date, n_obs, fill = "Reported")) +
  ylab("Number Fatalities") +
  xlab("Date") +
  scale_x_date(breaks = as.Date(c(
    "2022-01-01", "2022-01-07", "2022-01-13", "2022-01-19",
    "2022-01-26", "2022-02-01"
  )), date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.position = "bottom",
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-10, -10, -10, -10)
  ) +
  scale_fill_manual(values = c("Reported" = "grey30", "Occurred but not yet reported" = "gray"))

obs_plot

# ggsave(paste0("../plots/obs_", now, ".png"), units="in", dpi = 300, obs_plot, height = 3.5, width = 5.2)
ggsave(paste0("../plots/fig1.png"), units = "in", dpi = 300, obs_plot, height = 3.5, width = 5.2)
ggsave(paste0("../plots/fig1.tiff"), units = "in", dpi = 300, obs_plot, height = 3.5, width = 5.2, compression = "lzw")


# Fig 2
FHM_ICU <- read_excel("../data/FoHM/FHM_latest.xlsx",
  sheet = "Antal intensivvårdade per dag"
)

FHM_deaths <- read_excel("../data/FoHM/FHM_latest.xlsx",
  sheet = "Antal avlidna per dag", col_types = c("date", "numeric")
)

FHM_cases <- read_excel("../data/FoHM/FHM_latest.xlsx",
  sheet = "Antal per dag region"
)

# remove final row
FHM_deaths <- FHM_deaths[-nrow(FHM_deaths), ]

# Save dates
now <- ymd("2021-12-31")
start <- now - 180

# Create time series data frame
ts <- FHM_deaths %>%
  select(date = Datum_avliden, n_deaths = Antal_avlidna) %>%
  left_join(FHM_ICU %>% select(date = Datum_vårdstart, n_icu = Antal_intensivvårdade)) %>%
  left_join(FHM_cases %>% select(date = Statistikdatum, n_cases = Totalt_antal_fall)) %>%
  mutate(
    n_deaths_lag1 = lag(n_deaths, 1),
    mean_c = rollmean(n_cases, k = 21, fill = NA, align = "center"),
    mean_d = rollmean(n_deaths, k = 21, fill = NA, align = "center"),
    mean_i = rollmean(n_icu, k = 21, fill = NA, align = "center")
  )

FHM_ICU %>%
  select(date = Datum_vårdstart, n_icu = Antal_intensivvårdade) %>%
  filter(date >= "2020-10-20", date <= "2021-05-21") %>%
  summarise(n = sum(n_icu))

FHM_cases %>%
  select(date = Statistikdatum, n_cases = Totalt_antal_fall) %>%
  filter(date >= "2020-10-20", date <= "2021-05-21") %>%
  summarise(n = sum(n_cases))

max_death <- ts %>% filter(date < "2021-02-21") %>% 
  filter(mean_d == max(mean_d, na.rm = T))
max_icu <- ts %>% filter(date < "2021-02-21", date > "2020-11-01") %>% 
  filter(mean_i == max(mean_i, na.rm = T))
max_cases <- ts %>% filter( date < "2021-02-21") %>% 
  filter(mean_c == max(mean_c, na.rm = T))


timeseries_plots <- ts %>%
  filter(date > "2020-10-20", date <= "2021-05-21") %>%
  mutate(mean_c = mean_c / max_cases$mean_c, mean_i = mean_i / max_icu$mean_i, mean_d = mean_d / max_death$mean_d) %>%
  pivot_longer(c(mean_c, mean_d, mean_i)) %>%
  ggplot() +
  geom_segment(aes(
    x = as.Date("2020-12-27"), xend = as.Date("2020-12-27"),
    y = -Inf, yend = Inf
  ), color = "black") +
  geom_line(aes(as.Date(date), value, color = name, linetype = name)) +
  ylab("3-week average value (scaled)") +
  xlab("Date") +
  coord_cartesian(ylim = c(0.05, 1.15)) +
  annotate("text",
    x = as.Date("2020-12-14"), y = 0.2, cex = 2.1,
    label = "Vaccination\nstart 20-12-27", color = "black"
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-10, -10, -10, -10)
  ) +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  scale_y_continuous(breaks=seq(0, 1, by = 0.25))+
  scale_color_manual(
    values = c(mean_c = cols[6], mean_i = cols[3], mean_d = cols[1]),
    labels = c(mean_c = "Reported cases", mean_i = "ICU admissions", mean_d = "Fatalities")
  ) +
  scale_linetype_manual(
    values = c(mean_c = 1, mean_i = 4, mean_d = 2),
    labels = c(mean_c = "Reported cases", mean_i = "ICU admissions", mean_d = "Fatalities")
  )


timeseries_plots

# ggsave(paste0("../plots/ts.png"), timeseries_plots, width = 7,  height = 4)
ggsave(paste0("../plots/fig2.png"), units = "in", dpi = 300, timeseries_plots, height = 3.1, width = 5.2)
ggsave(paste0("../plots/fig2.tiff"), units = "in", dpi = 300, timeseries_plots, height = 3.1, width = 5.2, compression = "lzw")

# Fig 3 is a Tikz figure

# Fig 4

## Single reporting day
# Restrict dataset to a specific nowcast date
rep_dates <- list.files(path = paste0("../data/FoHM/")) %>%
  str_extract("\\d+-\\d+-\\d+") %>%
  as.data.frame() %>%
  distinct() %>%
  filter(. >= "2020-09-15") %>%
  t() %>%
  as.vector()

n <- 60
now <- ymd(rep_dates[n])
start <- ymd(now - 7 * 8 + 1)
dat_mod <- dat %>%
  filter(rep_date <= now)

N_a <- lapply(paste0("../results/N/", "N_mod_a_ph_2020-12-30.csv"), read_csv) %>% as.data.frame()
N_b <- lapply(paste0("../results/N/", "N_mod_b_cp_2020-12-30.csv"), read_csv) %>% as.data.frame()
N_d <- lapply(paste0("../results/N/", "N_mod_d_new2_2020-12-30.csv"), read_csv) %>% as.data.frame()

post_N_a <- tibble(
  date = seq(now - (55), now, "1 day"),
  med_a = apply(N_a, 2, median),
  q5_a = apply(N_a, 2, function(x) quantile(x, .025)),
  q95_a = apply(N_a, 2, function(x) quantile(x, .975))
)
post_N_b <- tibble(
  date = seq(now - (55), now, "1 day"),
  med_b = apply(N_b, 2, median),
  q5_b = apply(N_b, 2, function(x) quantile(x, .025)),
  q95_b = apply(N_b, 2, function(x) quantile(x, .975))
)

post_N_d <- tibble(
  date = seq(now - (55), now, "1 day"),
  med_d = apply(N_d, 2, median),
  q5_d = apply(N_d, 2, function(x) quantile(x, .025)),
  q95_d = apply(N_d, 2, function(x) quantile(x, .975))
)

dates <- c(seq(as.IDate("2020-11-25"), as.IDate("2020-12-30"), 12), as.Date("2020-12-30"))

snap_res_a <- dat_mod %>%
  group_by(date = death_date) %>%
  summarise(n_obs = n()) %>%
  right_join(tibble(date = seq(start, now, by = "1 day"))) %>%
  left_join(post_N_a) %>%
  left_join(dat %>% group_by(date = death_date) %>%
    summarise(n_true_retro = n())) %>%
  mutate_if(is.integer, ~ replace(., is.na(.), 0)) %>%
  filter(date > (now - 36)) %>%
  pivot_longer(c(med_a, n_true_retro)) %>%
  ggplot() +
  geom_line(aes(date, value, color = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_a, ymax = q95_a), fill = cols[4], alpha = .2) +
  geom_col(aes(date, n_obs / 2)) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 160)) +
  scale_color_manual(
    values = c(med_a = cols[4], n_true_retro = cols[1]),
    labels = c("R", "True number")
  ) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  scale_linetype_manual(
    values = c(med_a = 1, n_true_retro = 2),
    labels = c("R", "True number")
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-20, -15, -15, -15),
    legend.direction = "vertical",
    legend.spacing = unit(0.01, "mm"),
    plot.margin = margin(0.1, 0.2, 0.3, 0.1, "cm")
  )

snap_res_a


snap_res_b <- dat_mod %>%
  group_by(date = death_date) %>%
  summarise(n_obs = n()) %>%
  right_join(tibble(date = seq(start, now, by = "1 day"))) %>%
  left_join(post_N_b) %>%
  left_join(dat %>% group_by(date = death_date) %>%
    summarise(n_true_retro = n())) %>%
  mutate_if(is.integer, ~ replace(., is.na(.), 0)) %>%
  filter(date > (now - 36)) %>%
  pivot_longer(c(med_b, n_true_retro)) %>%
  ggplot() +
  geom_line(aes(date, value, col = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_b, ymax = q95_b), fill = cols[2], alpha = .2) +
  geom_col(aes(date, n_obs / 2)) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 160)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-20, -15, -15, -15),
    plot.margin = margin(0.1, 0.2, 0.3, 0.1, "cm"),
    legend.direction = "vertical",
    legend.spacing = unit(0.01, "mm")
  ) +
  scale_color_manual(
    values = c(med_b = cols[2], n_true_retro = cols[1]),
    labels = c("L(ICU)", "True number")
  ) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  scale_linetype_manual(
    values = c(med_b = 1, n_true_retro = 2),
    labels = c("L(ICU)", "True number")
  )

snap_res_b

snap_res_d <- dat_mod %>%
  group_by(date = death_date) %>%
  summarise(n_obs = n()) %>%
  right_join(tibble(date = seq(start, now, by = "1 day"))) %>%
  left_join(post_N_d) %>%
  left_join(dat %>% group_by(date = death_date) %>%
    summarise(n_true_retro = n())) %>%
  mutate_if(is.integer, ~ replace(., is.na(.), 0)) %>%
  filter(date > (now - 36)) %>%
  pivot_longer(c(med_d, n_true_retro)) %>%
  ggplot() +
  geom_line(aes(date, value, col = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_d, ymax = q95_d), fill = cols[8], alpha = .2) +
  geom_col(aes(date, n_obs / 2)) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 160)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-20, -15, -15, -15),
    plot.margin = margin(0.1, 0.2, 0.3, 0.1, "cm"),
    legend.direction = "vertical",
    legend.spacing = unit(0.01, "mm")
  ) +
  scale_color_manual(
    values = c(med_d = cols[8], n_true_retro = cols[1]),
    labels = c("RL(ICU)", "True number")
  ) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  scale_linetype_manual(
    values = c(med_d = 1, n_true_retro = 2),
    labels = c("RL(ICU)", "True number")
  )

snap_res_d

# Q-plots
p_est_a_1230 <- read_csv("../results/p/p_mod_a_ph_2020-12-30.csv")
p_est_d_1230 <- read_csv("../results/p/p_mod_d_new2_2020-12-30.csv")
p_est_b_1230 <- read_csv("../results/p/p_mod_b_cp_2020-12-30.csv")
p_1230_emp <- dat %>%
  mutate(delay = as.numeric(rep_date - death_date)) %>%
  filter(death_date >= "2020-11-01", death_date <= "2020-12-30") %>%
  select(death_date, delay) %>%
  group_by(death_date, delay) %>%
  summarise(n = n()) %>%
  right_join(tibble(expand.grid(
    death_date = seq(as.Date("2020-11-25"), as.Date("2020-12-30"), "1 day"),
    delay = 0:500
  ))) %>%
  mutate(n = replace_na(n, 0)) %>%
  arrange(death_date, delay) %>%
  mutate(frac = n / sum(n), cum_frac = cumsum(frac))

# Quantile plot A
## Quantile plot

q50_emp_30 <- p_1230_emp %>%
  filter(cum_frac >= 0.5) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_5_emp = delay)

q05_emp_30 <- p_1230_emp %>%
  filter(cum_frac >= 0.05) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_05_emp = delay)

q95_emp_30 <- p_1230_emp %>%
  filter(cum_frac >= 0.95) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_95_emp = delay)

p_1230_a_est <- p_est_a_1230 %>%
  pivot_longer(starts_with("p")) %>%
  group_by(name) %>%
  summarise(med = mean(value)) %>%
  mutate(
    day = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][2])),
    delay = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][3])) - 1
  ) %>%
  arrange(day, delay) %>%
  mutate(death_date = ymd("2020-12-30") - (56 - day)) %>%
  select(death_date, delay, est_p = med) %>%
  group_by(death_date) %>%
  mutate(cum_p = cumsum(est_p))

q50_est_30 <- p_1230_a_est %>%
  filter(cum_p >= 0.5) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_5_est = delay)

q05_est_30 <- p_1230_a_est %>%
  filter(cum_p >= 0.05) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_05 = delay)

q95_est_30 <- p_1230_a_est %>%
  filter(cum_p >= 0.95) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_95 = delay)

q_plot <- q50_emp_30 %>%
  left_join(q05_emp_30) %>%
  left_join(q95_emp_30) %>%
  left_join(q50_est_30) %>%
  left_join(q05_est_30) %>%
  left_join(q95_est_30) %>%
  pivot_longer(c(q_5_emp:q_95)) %>%
  mutate(name = factor(name, levels = c("q_05", "q_05_emp", "q_5_est", "q_5_emp", "q_95", "q_95_emp")))

q_plot$name %>% levels()

q_plot_a_30 <- q_plot %>%
  ggplot() +
  geom_line(aes(x = death_date, y = value, color = name, linetype = name)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  ylim(0, 35) +
  ylab("Delay (days)") +
  xlab("Date") +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    legend.spacing = unit(0, "cm"),
    legend.text.align = 0,
    plot.margin = margin(0.1, 0.4, 0.2, 0, "cm")
  ) +
  scale_color_manual(
    values = c(rep(c(cols[4], cols[1]), 3)),
    # breaks = c("q_05", "q_05_emp", "q_5_emp", "q_5_est","q_95", "q_95_emp"),
    labels = c(
      expression(paste(q[0.05] ~ R, " ")), expression(q[0.05] ~ Emp), expression(paste(q[0.50] ~ R, " ")),
      expression(q[0.50] ~ Emp), expression(paste(q[0.95] ~ R, " ")), expression(q[0.95] ~ Emp)
    )
  ) +
  scale_linetype_manual(
    values = c(4, 4, 1, 1, 3, 3),
    labels = c(
      expression(paste(q[0.05] ~ R, " ")), expression(q[0.05] ~ Emp), expression(paste(q[0.50] ~ R, " ")),
      expression(q[0.50] ~ Emp), expression(paste(q[0.95] ~ R, " ")), expression(q[0.95] ~ Emp)
    )
  )

q_plot_a_30


ggsave(paste0("../plots/q_plot_modR.png"), q_plot_a_30,
  width = 3.5,
  height = 3
)


# Quantil plot mod L
p_1230_b_est <- p_est_b_1230 %>%
  pivot_longer(starts_with("p")) %>%
  group_by(name) %>%
  summarise(med = mean(value)) %>%
  mutate(
    day = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][2])),
    delay = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][3])) - 1
  ) %>%
  arrange(day, delay) %>%
  mutate(death_date = ymd("2020-12-30") - (56 - day)) %>%
  select(death_date, delay, est_p = med) %>%
  group_by(death_date) %>%
  mutate(cum_p = cumsum(est_p))

q50_est_30 <- p_1230_b_est %>%
  filter(cum_p >= 0.5) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_5_est = delay)

q05_est_30 <- p_1230_b_est %>%
  filter(cum_p >= 0.05) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_05 = delay)

q95_est_30 <- p_1230_b_est %>%
  filter(cum_p >= 0.95) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_95 = delay)

q_plot <- q50_emp_30 %>%
  left_join(q05_emp_30) %>%
  left_join(q95_emp_30) %>%
  left_join(q50_est_30) %>%
  left_join(q05_est_30) %>%
  left_join(q95_est_30) %>%
  pivot_longer(c(q_5_emp:q_95)) %>%
  mutate(name = factor(name, levels = c("q_05", "q_05_emp", "q_5_est", "q_5_emp", "q_95", "q_95_emp")))

q_plot$name %>% levels()

q_plot_b_30 <- q_plot %>%
  ggplot() +
  geom_line(aes(x = death_date, y = value, color = name, linetype = name)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  ylim(0, 35) +
  ylab("Delay (days)") +
  xlab("Date") +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    legend.spacing = unit(0, "cm"),
    legend.text.align = 0,
    plot.margin = margin(0.1, 0.5, 0.2, 0, "cm")
  ) +
  scale_color_manual(
    values = c(rep(c(wes_cols[2], wes_cols[1]), 3)),
    labels = c(
      expression(paste(q[0.05] ~ L(ICU))), expression(paste(q[0.05] ~ Emp, " ")),
      expression(paste(q[0.50] ~ L(ICU))), expression(paste(q[0.50] ~ Emp, " ")),
      expression(paste(q[0.95] ~ L(ICU))), expression(paste(q[0.95] ~ Emp, " "))
    ),
    # breaks = c("q_05", "q_05_emp", "q_5_emp", "q_5_est","q_95", "q_95_emp")
  ) +
  scale_linetype_manual(
    values = c(4, 4, 1, 1, 3, 3),
    labels = c(
      expression(paste(q[0.05] ~ L(ICU))), expression(paste(q[0.05] ~ Emp, " ")),
      expression(paste(q[0.50] ~ L(ICU))), expression(paste(q[0.50] ~ Emp, " ")),
      expression(paste(q[0.95] ~ L(ICU))), expression(paste(q[0.95] ~ Emp, " "))
    )
  )

q_plot_b_30

ggsave(paste0("../plots/q_plot_modL.png"), q_plot_b_30,
  width = 3.5,
  height = 3
)


# Quantil plot mod RL
p_1230_d_est <- p_est_d_1230 %>%
  pivot_longer(starts_with("p")) %>%
  group_by(name) %>%
  summarise(med = mean(value)) %>%
  mutate(
    day = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][2])),
    delay = as.numeric(sapply(name, function(x) strsplit(x, "\\.")[[1]][3])) - 1
  ) %>%
  arrange(day, delay) %>%
  mutate(death_date = ymd("2020-12-30") - (56 - day)) %>%
  select(death_date, delay, est_p = med) %>%
  group_by(death_date) %>%
  mutate(cum_p = cumsum(est_p))


q50_est_30 <- p_1230_d_est %>%
  filter(cum_p >= 0.5) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_5_est = delay)

q05_est_30 <- p_1230_d_est %>%
  filter(cum_p >= 0.05) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_05 = delay)

q95_est_30 <- p_1230_d_est %>%
  filter(cum_p >= 0.95) %>%
  group_by(death_date) %>%
  filter(delay == min(delay)) %>%
  select(death_date, q_95 = delay)

q_plot_d <- q50_emp_30 %>%
  left_join(q05_emp_30) %>%
  left_join(q95_emp_30) %>%
  left_join(q50_est_30) %>%
  left_join(q05_est_30) %>%
  left_join(q95_est_30) %>%
  pivot_longer(c(q_5_emp:q_95)) %>%
  mutate(name = factor(name, levels = c("q_05", "q_05_emp", "q_5_est", "q_5_emp", "q_95", "q_95_emp")))

q_plot$name %>% levels()

q_plot_d_30 <- q_plot_d %>%
  ggplot() +
  geom_line(aes(x = death_date, y = value, color = name, linetype = name)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  ylim(0, 35) +
  ylab("Delay (days)") +
  xlab("Date") +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    legend.spacing = unit(0, "cm"),
    legend.text.align = 0,
    plot.margin = margin(0.1, 0.7, 0.2, 0, "cm")
  ) +
  scale_color_manual(
    values = c(rep(c(cols[8], cols[1]), 3)),
    # breaks = c("q_05", "q_05_emp", "q_5_emp", "q_5_est","q_95", "q_95_emp"),
    labels = c(
      q_05 = expression(paste(q[0.05] ~ RL(ICU))), q_05_emp = expression(paste(q[0.05] ~ Emp, " ")),
      q_5_est = expression(paste(q[0.50] ~ RL(ICU))), q_5_emp = expression(paste(q[0.50] ~ Emp, " ")),
      q_95 = expression(paste(q[0.95] ~ RL(ICU))), q_5_emp = expression(paste(q[0.95] ~ Emp, " "))
    )
  ) +
  scale_linetype_manual(
    values = c(4, 4, 1, 1, 3, 3),
    # breaks = c("q_05", "q_05_emp", "q_5_emp", "q_5_est","q_95", "q_95_emp"),
    labels = c(
      q_05 = expression(paste(q[0.05] ~ RL(ICU))), q_05_emp = expression(paste(q[0.05] ~ Emp, " ")),
      q_5_est = expression(paste(q[0.50] ~ RL(ICU))), q_5_emp = expression(paste(q[0.50] ~ Emp, " ")),
      q_95 = expression(paste(q[0.95] ~ RL(ICU))), q_5_emp = expression(paste(q[0.95] ~ Emp, " "))
    )
  )


q_plot_d_30

fig4 <- ggarrange(snap_res_a, q_plot_a_30, snap_res_b, q_plot_b_30, snap_res_d, q_plot_d_30, nrow = 3, ncol = 2)
fig4
ggsave(paste0("../plots/fig4.png"), units = "in", dpi = 300, fig4, height = 6.8, width = 5.2)
ggsave(paste0("../plots/fig4.tiff"), units = "in", dpi = 300, fig4, height = 6.8, width = 5.2)
ggsave(paste0("../plots/q_plot_modRL.png"), q_plot_d_30,
  width = 3.5,
  height = 3
)


# Fig 5
rep_plot_a <- res_df %>%
  pivot_longer(c(med_a, n_true_retro)) %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = value, color = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_a, ymax = q95_a), fill = cols[4], alpha = .2) +
  # geom_ribbon(aes(date, ymin=q5_d, ymax=q95_d), fill=wes_cols[5], alpha=.2) +
  ylab("Number Fatalities") +
  xlab("Date") +
  scale_x_date(date_breaks = "1 month", date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  coord_cartesian(ylim = c(0, 250)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(
    values = c(cols[4], cols[1]),
    labels = c("R", "True number")
  ) +
  scale_linetype_manual(
    values = c(1, 2),
    labels = c("R", "True number")
  )
rep_plot_a

rep_plot_b <- res_df %>%
  pivot_longer(c(med_b, n_true_retro)) %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = value, color = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_b, ymax = q95_b), fill = wes_cols[2], alpha = .2) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 250)) +
  scale_x_date(
    date_breaks = "1 month",
    date_labels = "%y-%m-%d",
    expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(
    values = c(med_b = cols[2], n_true_retro = cols[1]),
    labels = c("L(ICU)", "True number")
  ) +
  scale_linetype_manual(
    values = c(med_b = 1, n_true_retro = 2),
    labels = c("L(ICU)", "True number")
  )
rep_plot_b

rep_plot_d <- res_df %>%
  pivot_longer(c(med_d, n_true_retro)) %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = value, color = name, linetype = name)) +
  geom_ribbon(aes(date, ymin = q5_d, ymax = q95_d), fill = wes_cols[8], alpha = .2) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 250)) +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(
    values = c(cols[8], cols[1]),
    labels = c("RL(ICU)", "True number")
  ) +
  scale_linetype_manual(
    values = c(1, 2),
    labels = c("RL(ICU)", "True number")
  )
rep_plot_d

fig6 <- rep_plot_a + rep_plot_b + rep_plot_d +
  plot_annotation(tag_levels = c("A", "B")) +
  plot_layout(ncol = 1) & theme(
  legend.position = "bottom",
  plot.margin = margin(0.1, 0.1, 0.2, 0.1, "cm"),
  text = element_text(size = 8, family = "TT Arial"),
  legend.box.margin = margin(-15, -15, -15, -15)
)
fig6
ggsave(paste0("../plots/fig6.png"), units = "in", dpi = 300, fig6, height = 5.8, width = 5.2)
ggsave(paste0("../plots/fig6.tiff"), units = "in", dpi = 300, fig6, height = 5.8, width = 5.2)
ggsave(paste0("../plots/eval_period.png"), q_plot_d_30,
  width = 3.5,
  height = 3
)

# Fig 7
dates <- c(as.Date("2020-11-01"), as.Date("2021-01-01"), as.Date("2021-03-01"), as.Date("2021-05-01"))
# Log Score 7
log_plot_7 <- res_df %>%
  select(date, log_a_7, log_b_7, log_d_7) %>%
  pivot_longer(starts_with("log"), names_to = "model", values_to = "log") %>%
  # mutate(model = factor(model, levels = c("log_a", "log_d", "log_b"))) %>%
  ggplot(aes(x = date, y = log)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("LogS") +
  xlab("Date") +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(log_a_7 = "R", log_d_7 = "RL(ICU)", log_b_7 = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(log_a_7 = "R", log_d_7 = "RL(ICU)", log_b_7 = "L(ICU)")
  )

log_plot_7


crps_plot_7 <- res_df %>%
  select(date, crps_a_7, crps_b_7, crps_d_7) %>%
  pivot_longer(starts_with("crps"), names_to = "model", values_to = "crps") %>%
  ggplot(aes(x = date, y = crps)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("CRPS") +
  xlab("Date") +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(crps_a_7 = "R", crps_d_7 = "RL(ICU)", crps_b_7 = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(crps_a_7 = "R", crps_d_7 = "RL(ICU)", crps_b_7 = "L(ICU)")
  )

crps_plot_7

fig7 <- log_plot_7 + crps_plot_7 + plot_annotation(tag_levels = c("A", "B")) +
  plot_layout(guides = "collect", ncol = 1) & theme(
  legend.position = "bottom",
  plot.margin = margin(0.1, 0.1, 0.1, 0.1, "cm"),
  text = element_text(size = 8, family = "TT Arial"),
  legend.box.margin = margin(-15, -15, -15, -15)
)


fig7

ggsave(paste0("../plots/fig7.png"), units = "in", dpi = 300, fig7, height = 3.5, width = 5.2)
ggsave(paste0("../plots/fig7.tiff"), units = "in", dpi = 300, fig7, height = 3.5, width = 5.2)
ggsave(paste0("../plots/scores.png"), fig6,
  width = 3.5,
  height = 3
)

# SI
# Fig S1

dates <- c(seq(as.IDate("2020-11-25"), as.IDate("2020-12-30"), 12), as.Date("2020-12-30"))

plot_est_a_1230 <- p_1230_a_est %>%
  filter(delay %in% c(1, 3, 7, 10, 14, 21, 35)) %>%
  mutate(Delay = factor(delay,
    levels = rev(c(1, 3, 7, 10, 14, 21, 35)),
    labels = rev(c(as.character(c(1, 3, 7, 10, 14, 21)), ">21"))
  )) %>%
  filter(death_date >= "2020-11-25") %>%
  ggplot() +
  geom_ribbon(aes(death_date, ymax = cum_p, ymin = 0, fill = Delay)) +
  geom_line(aes(death_date, y = cum_p, group = Delay)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d") +
  ylab("Estimated probability R") +
  xlab("Date") +
  scale_fill_manual(values = pal) +
  theme( # legend.background = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.margin = margin(),
    text = element_text(size = 8, family = "TT Arial")
    #  legend.title = "element_blank()"
  ) +
  guides(fill = guide_legend(nrow = 1, byrow = TRUE))


plot_est_a_1230


plot_est_b_1230 <- p_1230_b_est %>%
  filter(delay %in% c(1, 3, 7, 10, 14, 21, 35)) %>%
  mutate(Delay = factor(delay,
    levels = rev(c(1, 3, 7, 10, 14, 21, 35)),
    labels = rev(c(as.character(c(1, 3, 7, 10, 14, 21)), ">21"))
  )) %>%
  filter(death_date >= "2020-11-25") %>%
  ggplot() +
  geom_ribbon(aes(death_date, ymax = cum_p, ymin = 0, fill = Delay)) +
  geom_line(aes(death_date, y = cum_p, group = Delay)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d") +
  ylab("Estimated probability L(ICU)") +
  xlab("Date") +
  scale_fill_manual(values = pal) +
  theme( # legend.background = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.margin = margin(),
    text = element_text(size = 8, family = "TT Arial")
    #  legend.title = "element_blank()"
  ) +
  guides(fill = guide_legend(nrow = 1, byrow = TRUE))

plot_est_b_1230 <- p_1230_b_est %>%
  filter(delay %in% c(1, 3, 7, 10, 14, 21, 35)) %>%
  mutate(Delay = factor(delay,
    levels = rev(c(1, 3, 7, 10, 14, 21, 35)),
    labels = rev(c(as.character(c(1, 3, 7, 10, 14, 21)), ">21"))
  )) %>%
  filter(death_date >= "2020-11-25") %>%
  ggplot() +
  geom_ribbon(aes(death_date, ymax = cum_p, ymin = 0, fill = Delay)) +
  geom_line(aes(death_date, y = cum_p, group = Delay)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d") +
  ylab("Estimated probability L(ICU)") +
  xlab("Date") +
  scale_fill_manual(values = pal) +
  theme( # legend.background = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.margin = margin(),
    text = element_text(size = 8, family = "TT Arial")
    #  legend.title = "element_blank()"
  ) +
  guides(fill = guide_legend(nrow = 1, byrow = TRUE))

plot_est_d_1230 <- p_1230_d_est %>%
  filter(delay %in% c(1, 3, 7, 10, 14, 21, 35)) %>%
  mutate(Delay = factor(delay,
    levels = rev(c(1, 3, 7, 10, 14, 21, 35)),
    labels = rev(c(as.character(c(1, 3, 7, 10, 14, 21)), ">21"))
  )) %>%
  filter(death_date >= "2020-11-25") %>%
  ggplot() +
  geom_ribbon(aes(death_date, ymax = cum_p, ymin = 0, fill = Delay)) +
  geom_line(aes(death_date, y = cum_p, group = Delay)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d") +
  ylab("Estimated probability RL(ICU)") +
  xlab("Date") +
  scale_fill_manual(values = pal) +
  theme( # legend.background = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.margin = margin(),
    text = element_text(size = 8, family = "TT Arial")
    #  legend.title = "element_blank()"
  ) +
  guides(fill = guide_legend(nrow = 1, byrow = TRUE))


plot_emp_1230 <- p_1230_emp %>%
  filter(delay %in% c(1, 3, 7, 10, 14, 21, 500)) %>%
  mutate(Delay = factor(delay,
    levels = rev(c(1, 3, 7, 10, 14, 21, 500)),
    labels = rev(c(as.character(c(1, 3, 7, 10, 14, 21)), ">21"))
  )) %>%
  ggplot() +
  geom_ribbon(aes(death_date, ymax = cum_frac, ymin = 0, fill = Delay)) +
  geom_line(aes(death_date, y = cum_frac, group = Delay)) +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d") +
  ylab("Empirical probability") +
  xlab("Date") +
  scale_fill_manual(values = pal) +
  theme( # legend.background = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.margin = margin(),
    text = element_text(size = 8, family = "TT Arial")
    #  legend.title = "element_blank()"
  ) +
  guides(fill = guide_legend(nrow = 1, byrow = TRUE))


figS1 <- plot_emp_1230 + plot_est_a_1230 +
  plot_est_b_1230 + plot_est_d_1230 +
  plot_layout(guides = "collect", ncol = 2) & theme(
  legend.position = "bottom",
  plot.margin = margin(0.1, 0.2, 0.3, 0.1, "cm"),
  text = element_text(size = 8, family = "TT Arial"),
  legend.box.margin = margin(-5, -10, -10, -10)
)


ggsave(paste0("../plots/figS1.png"), units = "in", dpi = 300, figS1, height = 4, width = 5.2)
ggsave(paste0("../plots/figS1.tiff"), units = "in", dpi = 300, figS1, height = 4, width = 5.2)


# S2
rep_plot_dc <- res_df %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = med_d_c, color = "RL(Cases)")) +
  geom_line(aes(y = n_true_retro, color = "True number"), lty = 2) +
  geom_ribbon(aes(date, ymin = q5_d_c, ymax = q95_d_c), fill = wes_cols[8], alpha = .2) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 250)) +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(
    values = c(wes_cols[8], cols[1]),
    labels = c("RL(Cases)", "True number")
  ) +
  scale_linetype_manual(
    values = c(1, 2),
    labels = c("RL(Cases)", "True number")
  )
rep_plot_dc

rep_plot_dic <- res_df %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = med_e, color = "RL(ICU, Cases)")) +
  geom_line(aes(y = n_true_retro, color = "True number"), lty = 2) +
  geom_ribbon(aes(date, ymin = q5_e, ymax = q95_e), fill = wes_cols[9], alpha = .2) +
  ylab("Number Fatalities") +
  xlab("Date") +
  coord_cartesian(ylim = c(0, 250)) +
  scale_x_date(
    date_breaks = "1 month", date_labels = "%y-%m-%d",
    limits = c(as.Date("2020-10-20"), as.Date("2021-05-21")), expand = c(0.02, 0.02)
  ) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(
    values = c(wes_cols[9], cols[1]),
    labels = c("RL(ICU, Cases)", "True number")
  ) +
  scale_linetype_manual(
    values = c(1, 2),
    labels = c("RL(ICU, Cases)", "True number")
  )
rep_plot_dic

figS2 <- ggarrange(rep_plot_dc, rep_plot_dic, nrow = 2)
figS2
ggsave(paste0("../plots/figS2.png"), units = "in", dpi = 300, figS2, height = 3.9, width = 5.2)
ggsave(paste0("../plots/figS2.tiff"), units = "in", dpi = 300, figS2, height = 3.9, width = 5.2)

# Fig S3 (beta)


# Beta 0
res_beta_0 <- read_csv("../results/results_beta_0.csv") %>% filter(date >= "2020-10-15", date <= "2021-05-14")
beta_0_plot <- res_beta_0 %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = med_b, color = "L(ICU)")) +
  geom_ribbon(aes(date, ymin = q5_b, ymax = q95_b), fill = wes_cols[14], alpha = .2) +
  ylab(expression(beta[0])) +
  xlab("Date") +
  scale_x_date(date_breaks = "1 month", date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(values = c("L(ICU)" = wes_cols[14]))
beta_0_plot

# Beta 1
res_beta_1 <- read_csv("../results/results_beta_1_mod_b.csv") %>% filter(date >= "2020-10-15", date <= "2021-05-14")
beta_1_plot <- res_beta_1 %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = med_b, color = "L(ICU)")) +
  geom_ribbon(aes(date, ymin = q5_b, ymax = q95_b), fill = wes_cols[13], alpha = .2) +
  ylab(expression(beta[1])) +
  xlab("Date") +
  scale_x_date(date_breaks = "1 month", date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(values = c("L(ICU)" = wes_cols[13]))
beta_1_plot

figS3 <- ggarrange(beta_0_plot, beta_1_plot, nrow = 2)
figS3
ggsave(paste0("../plots/figS3.png"), units = "in", dpi = 300, figS3, height = 3.9, width = 5.2)
ggsave(paste0("../plots/figS3.tiff"), units = "in", dpi = 300, figS3, height = 3.9, width = 5.2)



res_beta_1_d <- read_csv("../results/results_beta_1_mod_d.csv") %>% filter(date >= "2020-10-15", date <= "2021-05-15")
figS4 <- res_beta_1_d %>%
  ggplot(aes(x = date)) +
  geom_line(aes(y = med_d, color = "RL(ICU)")) +
  geom_ribbon(aes(date, ymin = q5_d, ymax = q95_d), fill = wes_cols[3], alpha = .2) +
  ylab(expression(beta[1])) +
  xlab("Date") +
  scale_x_date(date_breaks = "1 month", expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15)
  ) +
  scale_color_manual(values = c("RL(ICU)" = wes_cols[3]))
figS4

ggsave(paste0("../plots/figS4.png"), units = "in", dpi = 300, figS4, height = 2, width = 5.2)
ggsave(paste0("../plots/figS4.tiff"), units = "in", dpi = 300, figS4, height = 2, width = 5.2)


# SI Fig1 RMSE

# RMSE 7days
# dates <- c(as.Date("2020-11-01"), as.Date("2021-01-01"), as.Date("2021-03-01"), as.Date("2021-05-01"))

rmse7_plot <- res_df %>%
  select(date, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("err"), names_to = "model", values_to = "rmse") %>%
  ggplot(aes(x = date, y = rmse)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("RMSE") +
  xlab("Date") +
  scale_x_date(breaks = dates, date_labels = "%y-%m-%d", expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0.5, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(err_a_7 = "R", err_d_7 = "RL(ICU)", err_b_7 = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(err_a_7 = "R", err_d_7 = "RL(ICU)", err_b_7 = "L(ICU)")
  )

rmse7_plot

ggsave(paste0("../plots/S1_fig.png"), units = "in", dpi = 300, rmse7_plot, height = 1.75, width = 5.2)
ggsave(paste0("../plots/S1_fig.tiff"), units = "in", dpi = 300, rmse7_plot, height = 1.75, width = 5.2)

# S2 Fig
err_delay <- read_csv("../results/error_by_delay.csv") %>% as.data.frame()

rmse_plot <- err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("rmse"), names_to = "model", values_to = "rmse") %>%
  ggplot(aes(x = delay, y = rmse)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("RMSE") +
  xlab(expression(paste("Days since day ", italic("T")))) +
  scale_x_continuous(breaks = 0:7 * 5, expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0.5, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(rmse_a = "R", rmse_d = "RL(ICU)", rmse_b = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(rmse_a = "R", rmse_d = "RL(ICU)", rmse_b = "L(ICU)")
  )

rmse_plot

log_plot <- err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("log"), names_to = "model", values_to = "log") %>%
  ggplot(aes(x = delay, y = log)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("logS") +
  xlab(expression(paste("Days since day ", italic("T")))) +
  scale_x_continuous(breaks = 0:7 * 5, expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0.5, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(log_a = "R", log_d = "RL(ICU)", log_b = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(log_a = "R", log_d = "RL(ICU)", log_b = "L(ICU)")
  )

log_plot


crps_plot <- err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("crps"), names_to = "model", values_to = "crps") %>%
  ggplot(aes(x = delay, y = crps)) +
  geom_line(aes(color = model, linetype = model)) +
  ylab("CRPS") +
  xlab(expression(paste("Days since day ", italic("T")))) +
  scale_x_continuous(breaks = 0:7 * 5, expand = c(0.02, 0.02)) +
  theme(
    legend.background = element_blank(),
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15),
    plot.margin = margin(0.1, 0.1, 0.5, 0, "cm")
  ) +
  scale_color_manual(
    values = c(cols[4], cols[2], cols[8]),
    labels = c(crps_a = "R", crps_d = "RL(ICU)", crps_b = "L(ICU)")
  ) +
  scale_linetype_manual(
    values = c(1, 3, 2),
    labels = c(crps_a = "R", crps_d = "RL(ICU)", crps_b = "L(ICU)")
  )

crps_plot

figS2 <- crps_plot + log_plot +
  rmse_plot +
  plot_annotation(tag_levels = c("A", "B")) +
  plot_layout(guides = "collect", ncol = 3) & theme(
  legend.position = "bottom",
  plot.margin = margin(0.1, 0.1, 0.1, 0.1, "cm"),
  text = element_text(size = 8, family = "TT Arial"),
  legend.box.margin = margin(-15, -15, -15, -15)
)

figS2
ggsave(paste0("../plots/figS2.png"), units = "in", dpi = 300, figS2, height = 2.5, width = 5.2)
ggsave(paste0("../plots/figS2.tiff"), units = "in", dpi = 300, figS2, height = 2.5, width = 5.2)


# Fig S2
# Transforming (reversing) dates
c_trans <- function(a, b, breaks = b$breaks, format = b$format) {
  a <- as.trans(a)
  b <- as.trans(b)
  
  name <- paste(a$name, b$name, sep = "-")
  
  trans <- function(x) a$trans(b$trans(x))
  inv <- function(x) b$inverse(a$inverse(x))
  
  trans_new(name, trans, inverse = inv, breaks = breaks, format=format)
  
}

D <- 35
now <- "2020-12-30" %>% as.Date()
start <- now -35

plot_dat <- dat %>%
  group_by(death_date, rep_date) %>%
  summarise(n_true_retro = as.factor(n())) %>%
  mutate(
    delay = rep_date - death_date,
    Reported = case_when(
      death_date <= now - delay ~ "n_obs",
      death_date > now - delay ~ "unrep"
    ),
    Reported = factor(Reported, levels= c("n_obs", "unrep"), ordered = T),
    death_date = as.POSIXct(death_date)
  ) %>% 
  filter(death_date >= start, death_date <= now, delay <= D)
rev_date <- c_trans("reverse", "time")

p1 <- plot_dat %>% ggplot(aes(x = as.numeric(delay), y = death_date)) +
  theme_bw() +
  xlab("Days delay") +
  ylab("Date") +
  geom_tile(aes(fill = Reported)) +
  theme(
    axis.ticks = element_blank(),
    axis.line = element_blank(),
    panel.border = element_blank(),
    panel.grid.major = element_line(color = "#eeeeee")
  ) +
  geom_text(aes(label = n_true_retro)) +
  scale_y_continuous(trans = rev_date, 
                     breaks = as.POSIXct(c("2020-11-25", "2020-12-02", "2020-12-09",
                                           "2020-12-16","2020-12-23","2020-12-30")))+
  scale_fill_manual(values = c(n_obs="green",unrep = "red" ),
                    labels = c("Reported", "Occurred but not yet reported"))
  #+
  #theme(legend.background = element_blank(),
   #     legend.title = element_blank(),
    #    legend.position = "none"
#  )
dat_mod <- dat %>% filter(rep_date <= now)
p2 <- dat %>% group_by(date=death_date) %>%
  summarise(n_true_retro=n()) %>%
  left_join(dat_mod %>% group_by(date=death_date) %>%
              summarise(n_obs=n())) %>%
  mutate_if(is.integer, ~replace(., is.na(.), 0)) %>%
  filter(date >= start, date <= now) %>%
  mutate(unrep = n_true_retro-n_obs) %>% 
  gather(Reported, n, c(n_obs,unrep)) %>% 
  mutate(Reported = factor(Reported, levels= c( "unrep", "n_obs"), ordered = T)) %>% 
  ggplot()+
  geom_col(aes(date, n, fill = Reported)) +
  xlab("Date") +
  ylab("Deaths") +
  scale_x_date(breaks = as.Date(c("2020-11-25", "2020-12-02", "2020-12-09",
                                  "2020-12-16","2020-12-23","2020-12-30")), date_labels = "%y-%m-%d")+
  scale_fill_manual(values = c(n_obs="green",unrep = "red" ),
                    labels = c("Reported", "Occurred but not yet reported"))+
  theme(legend.background = element_blank(),
        legend.title = element_blank(),
        legend.position = "bottom"
  )

p1 + p2 + plot_annotation(tag_levels = c("A", "B")) +
  plot_layout(guides = "collect", ncol = 3) & theme(
    legend.position = "bottom",
    plot.margin = margin(0.1, 0.1, 0.1, 0.1, "cm"),
    text = element_text(size = 8, family = "TT Arial"),
    legend.box.margin = margin(-15, -15, -15, -15))

# Table
err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("crps"), names_to = "model", values_to = "crps") %>%
  group_by(model) %>%
  filter(delay %in% 0:6) %>%
  summarise(sum(crps) / 7) %>%
  view()

err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("log"), names_to = "model", values_to = "log") %>%
  group_by(model) %>%
  filter(delay %in% 0:6) %>%
  summarise(sum(log) / 7)

err_delay %>% # select(delay, err_a_7, err_d_7, err_b_7) %>%
  pivot_longer(starts_with("rmse"), names_to = "model", values_to = "rmse") %>%
  group_by(model) %>%
  filter(delay %in% 0:6) %>%
  summarise(sum(rmse) / 7)
